requirements.txt
# Teltonika Tracking System Requirements
# Core Django and REST Framework
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1

# Database
psycopg2-binary==2.9.10

# Additional utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Production server (optional)
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools (optional)
django-extensions==3.2.3 